% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lGetDailyDataForTickersFromStooq}
\alias{lGetDailyDataForTickersFromStooq}
\title{Get daily data from Stooq}
\usage{
lGetDailyDataForTickersFromStooq(
  cTickers,
  dateStartDate,
  dateEndDate,
  bAddTickerToColnames = TRUE
)
}
\arguments{
\item{cTickers}{- character vector of ticker symbols;}

\item{dateStartDate}{- Date class scalar, points at the first day of the interval
for which the data should be fetched}

\item{dateEndDate}{- Date class scalar, points at the last day of the interval
for which the data should be fetched}

\item{bAddTickerToColnames}{- logical scalar, indicates whether the ticker
should be added as prefix to the non-Date columns names of the columns
of the data set retrieved from Stooq. Deault: TRUE}
}
\value{
list of data.tables; each member of the list corresponds to one ticker
symbol in input vector cTickers
}
\description{
Get daily time series from Stooq for indicated tickers
}
\details{
This function allows for fetching historical time series of DAILY prices for by tickers
from free online database at https://stooq.com/.
Remark re handling of the wrong ticker symbols passed inside cTickers:
}
