% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lGetDailyDataForTickersFromStooq}
\alias{lGetDailyDataForTickersFromStooq}
\title{Get daily data from Stooq}
\usage{
lGetDailyDataForTickersFromStooq(cTickers, dateStartDate, dateEndDate)
}
\arguments{
\item{cTickers}{- character vector of ticker symbols;}

\item{dateStartDate}{- Date class scalar, points at the first day of the interval
for which the data should be fetched}

\item{dateEndDate}{- Date class scalar, points at the last day of the interval
for which the data should be fetched}
}
\value{
list of data.tables; each member of the list corresponds to one ticker
symbol in input vector cTickers
}
\description{
Get daily time series from Stooq for indicated tickers
}
\details{
This function allows for fetching historical time series of DAILY prices for by tickers
from free online database at https://stooq.com/.
Remark re handling of the wrong ticker symbols passed inside cTickers:
}
